+++ CHANGELOG.md
@@ -2,4 +2,8 @@
 
 ## 0.1.0.0 -- 2024-04-02
 
-* Draft version. Return a random quote as json on.
+* Draft version.
+* Use a sqlite db to store quotes.
+* Use sqllite-simple to interact with the db.
+* Return a random quote as json data.
+